Update Calculator and Readme
Added a feature for calculating the degree classification for Cyber Security Modules using the modules for cohort 2022 students, at University of Staffordshire London. There is also a menu, which can be used to select the old calculator for finding out what a student need to get on average in 3rd year in order to get a 1st honours.